+++ src/pages/settings/settings.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { NavController } from 'ionic-angular';
 import { Logger } from '../../providers/logger/logger';
+
 import * as _ from 'lodash';
 
 // providers
